Add the position calculator
At this point we've got a basic 3D calibration mechanism, though now
based on polar coordinates. The angle scanner now turns off the LEDs at
the end of a scan (and generates the "scene" more efficiently). The
positions calculator accepts multiple angle scanners and interpolates
their guessed positions into a 3D model where each LED has a height from
the base, and angle around the trunk, and a radius from the trunk.

This should be relatively trivial to convert to X, Y, Z when we need it
but is easier to handle for interpolation purposes and should enable
some interesting animations too.

This commit does *not* include the necessary changes to the storage.