Finish no-JS implementation for preset creation
Also overhauled the JS implementation to be a bit simpler. The "Custom"
animation with a data box is now gone (never used it anyway), and the
caching of the animation data is done in a dataset on the form itself.
The "animations" and "presets" globals are now obtained from the server
on-demand (to avoid serving a load of JSON to browsers that aren't going
to use it anyway). Finally, fixed up non-click submission in the
creation form.